+++ comics/comics/heltnils.py
@@ -6,11 +6,10 @@ class ComicData(ComicDataBase):
     name = 'Helt Nils'
     language = 'no'
     url = 'http://www.pondus.no/'
+    active = False
     rights = 'Nils Ofstad'
 
 
 class Crawler(PondusNoCrawlerBase):
-    schedule = 'Mo,Fr'
-
     def crawl(self, pub_date):
-        return self.crawl_helper('20060')
+        pass  # Comic no longer published